requirements.txt
fastapi[all]
requests
pyjwt
boto3
sqlalchemy
pymysql
python-jose